refactor logging
