+++ human-dont-get-mad/src/game/Log.java
@@ -1,4 +1,4 @@
-package client;
+package game;
 
 public enum Log {
 error,
