+++ human-dont-get-mad/src/game/MsgType.java
@@ -1,4 +1,4 @@
-package client;
+package game;
 
 public enum MsgType {
 	register,
