+++ human-dont-get-mad/src/server/ClientThread.java
@@ -4,6 +4,8 @@ import java.io.IOException;
 import java.io.PrintWriter;
 import java.net.Socket;
 import java.util.Scanner;
+import game.Log;
+import game.LogController;
 
 public class ClientThread implements Runnable {
 	private Socket socket;
@@ -20,7 +22,7 @@ public class ClientThread implements Runnable {
 	@Override
 	public void run() {
 		//setup
-		Server.log(Log.info, "Client connected: " + socket);
+		LogController.log(Log.info, "Client connected: " + socket);
 		try {
 			in = new Scanner(socket.getInputStream());
 			out = new PrintWriter(socket.getOutputStream(), true);
@@ -30,15 +32,15 @@ public class ClientThread implements Runnable {
 				controller.deciverInput(this, in.nextLine());
 			}
 		}
-		catch (Exception e) { Server.log(Log.error, e.toString()); }
+		catch (Exception e) { LogController.log(Log.error, e.toString()); }
 		finally {
 			try {
 				in.close();
 				out.close();
 				socket.close();
 			}
-			catch (IOException e) { Server.log(Log.error, e.toString()); }
-			Server.log(Log.info, "Client disconnect: " + socket);
+			catch (IOException e) { LogController.log(Log.error, e.toString()); }
+			LogController.log(Log.info, "Client disconnect: " + socket);
 		}
 	}
