--- human-dont-get-mad/src/server/Log.java
@@ -1,7 +1,0 @@
-package server;
-
-public enum Log {
-error,
-info,
-debug
-}
