--- human-dont-get-mad/src/server/MsgType.java
@@ -1,14 +1,0 @@
-package server;
-
-public enum MsgType {
-	register,
-	ready,
-	move,
-	welcome,
-	assignColor,
-	update,
-	turn,
-	playerDisconnected,
-	message,
-	error
-}
